# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Patient {
  id: number;
  name: string;
  /** @nullable */
  ageYears: number | null;
  /** @nullable */
  weightKg: number | null;
  /** @nullable */
  sex: string | null;
  /** @nullable */
  village: string | null;
  /** @nullable */
  phoneNumber: string | null;
  createdAt: string;
}

export interface PatientInput {
  /** @minLength 1 */
  name: string;
  ageYears?: number;
  weightKg?: number;
  sex?: string;
  village?: string;
  phoneNumber?: string;
}

export interface PatientUpdate {
  /** @minLength 1 */
  name?: string;
  ageYears?: number;
  weightKg?: number;
  sex?: string;
  village?: string;
  phoneNumber?: string;
}

export interface Visit {
  id: number;
  patientId: number;
  symptoms: string;
  /** @nullable */
  photoBase64: string | null;
  /** routine | needs_monitoring | urgent_referral */
  urgency: string;
  conditions: string;
  recommendation: string;
  /** @nullable */
  dosage: string | null;
  dangerSignDetected: boolean;
  referralRequired: boolean;
  guidanceEnglish: string;
  guidanceYoruba: string;
  createdAt: string;
}

export interface PatientWithVisits {
  id: number;
  name: string;
  /** @nullable */
  ageYears: number | null;
  /** @nullable */
  weightKg: number | null;
  /** @nullable */
  sex: string | null;
  /** @nullable */
  village: string | null;
  /** @nullable */
  phoneNumber: string | null;
  createdAt: string;
  visits: Visit[];
}

export interface VisitInput {
  /** @minLength 1 */
  symptoms: string;
  /** Base64-encoded image (JPEG or PNG) */
  photoBase64?: string;
  /** Comma-separated list of medications to check dosage for */
  medications?: string;
}

export interface VisitWithPatient {
  id: number;
  patientId: number;
  patientName: string;
  symptoms: string;
  urgency: string;
  dangerSignDetected: boolean;
  referralRequired: boolean;
  guidanceEnglish: string;
  createdAt: string;
}

export interface DashboardStats {
  totalPatients: number;
  visitsToday: number;
  urgentCasesToday: number;
  visitsThisWeek: number;
  referralsThisWeek: number;
}

export interface ApiError {
  error: string;
}

export type ListPatientsParams = {
search?: string;
};