# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
}