# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalPatients: number;
  visitsToday: number;
  urgentCasesToday: number;
  visitsThisWeek: number;
  referralsThisWeek: number;
}