# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  /** @nullable */
  ageYears: number | null;
  /** @nullable */
  weightKg: number | null;
  /** @nullable */
  sex: string | null;
  /** @nullable */
  village: string | null;
  /** @nullable */
  phoneNumber: string | null;
  createdAt: Date;
}