# lib/api-zod/src/generated/types/patientInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface PatientInput {
  /** @minLength 1 */
  name: string;
  ageYears?: number;
  weightKg?: number;
  sex?: string;
  village?: string;
  phoneNumber?: string;
}