# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  patientId: number;
  symptoms: string;
  /** @nullable */
  photoBase64: string | null;
  /** routine | needs_monitoring | urgent_referral */
  urgency: string;
  conditions: string;
  recommendation: string;
  /** @nullable */
  dosage: string | null;
  dangerSignDetected: boolean;
  referralRequired: boolean;
  guidanceEnglish: string;
  guidanceYoruba: string;
  createdAt: Date;
}