# lib/api-zod/src/generated/types/visitInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface VisitInput {
  /** @minLength 1 */
  symptoms: string;
  /** Base64-encoded image (JPEG or PNG) */
  photoBase64?: string;
  /** Comma-separated list of medications to check dosage for */
  medications?: string;
}