# lib/api-zod/src/generated/types/visitWithPatient.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MamaCare Triage API — Maternal & Child Health Triage for Community Health Workers
 * OpenAPI spec version: 0.1.0
 */

export interface VisitWithPatient {
  id: number;
  patientId: number;
  patientName: string;
  symptoms: string;
  urgency: string;
  dangerSignDetected: boolean;
  referralRequired: boolean;
  guidanceEnglish: string;
  createdAt: Date;
}