# lib/db/src/schema/patients.ts
import { integer, numeric, pgTable, serial, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const patientsTable = pgTable("patients", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  ageYears: integer("age_years"),
  weightKg: numeric("weight_kg", { precision: 5, scale: 2 }),
  sex: text("sex"),
  village: text("village"),
  phoneNumber: text("phone_number"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertPatientSchema = createInsertSchema(patientsTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});

export type Patient = typeof patientsTable.$inferSelect;
export type InsertPatient = z.infer<typeof insertPatientSchema>;